fix migration: custom permission

+++ database/migrations/2025_04_07_114554_add_custom_permission_field_to_appfeatures.php
@@ -8,15 +8,19 @@ return new class extends Migration
 {
     public function up()
     {
-        Schema::table('appfeatures', function (Blueprint $table) {
-            $table->text('custom_permission')->nullable();
-        });
+        if (!Schema::hasColumn('appfeatures', 'custom_permission')) {
+            Schema::table('appfeatures', function (Blueprint $table) {
+                $table->text('custom_permission')->nullable();
+            });
+        }
     }
 
     public function down()
     {
-        Schema::table('appfeatures', function (Blueprint $table) {
-            $table->dropColumn('custom_permission');
-        });
+        if (Schema::hasColumn('appfeatures', 'custom_permission')) {
+            Schema::table('appfeatures', function (Blueprint $table) {
+                $table->dropColumn('custom_permission');
+            });
+        }
     }
 };
